backend/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.3
django-binary-database-files==1.0.18
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0